Adding router.emit func, new line in file
